pyproject.toml
[tool.poetry]
name = "uptime-monitor"
version = "0.1.0"
description = ""
authors = ["Anson <58594437+AnsonDev42@users.noreply.github.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.0.2"
psycopg = "^3.1.18"
httpx = "^0.27.0"
pydantic = "^2.6.2"
celery = "^5.3.6"
djangorestframework = "^3.14.0"
markdown = "^3.5.2"
django-filter = "^23.5"
requests = "^2.31.0"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
